fix cstyle renderer
